[HP][CRUDE, H2, JPA, 4 Layer] - model package renamed to dto, Javadoc updated in StudentService interface.
