+++ src/main/java/com/example/springboothelperproject/dto/Student.java
@@ -1,4 +1,4 @@
-package com.example.springboothelperproject.model;
+package com.example.springboothelperproject.dto;
 
 import com.example.springboothelperproject.enums.Department;
 import com.example.springboothelperproject.enums.Gender;
