+++ src/main/java/com/example/springboothelperproject/service/StudentServiceImpl.java
@@ -2,7 +2,7 @@ package com.example.springboothelperproject.service;
 
 import com.example.springboothelperproject.dao.StudentDao;
 import com.example.springboothelperproject.entity.StudentEntity;
-import com.example.springboothelperproject.model.Student;
+import com.example.springboothelperproject.dto.Student;
 import com.sun.jdi.request.InvalidRequestStateException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
